Move annotation check to separate module
